Map Header Info To CourseID
Used the same header that was used for the amount id in order to parse
the type of course. (header before quality indicator). Most likely a
database lookup for it would be cleaner. The mapping function is called
MapHeaderToCourseID and functions just like the MapHeaderToAmountID
function but throws a no such element exception.